requirements.txt
Flask==3.0.0
gunicorn==21.2.0
openai>=1.54.0
anthropic==0.8.0
python-dotenv==1.0.0
markdown2==2.4.10
APScheduler==3.10.4
psycopg2-binary==2.9.9

